feat(scripts/inject): better path auto-fix
- also update to deno v2

+++ scripts/inject/index.mts
@@ -3,9 +3,7 @@
 import { copy, ensureDir, exists } from 'std/fs/mod.ts';
 import { basename, join } from 'std/path/mod.ts';
 
-export async function injectLocal(location: string, type: 'packed' | 'unpacked' | 'symlink' = 'packed') {
-  if (!import.meta.dirname) throw new Error('This script must be run locally');
-
+async function setup(location: string) {
   const stat = await Deno.stat(location)
     .catch((e: unknown) => {
       if (e instanceof Deno.errors.NotFound) throw new Error(`"${location}" does not exist`);
@@ -18,25 +16,31 @@ export async function injectLocal(location: string, type: 'packed' | 'unpacked' 
     dir: `${location.replace(/\.asar$/, '').replace(/[\\/]$/, '')}-original`,
   };
 
-  const originalLocation = stat.isFile
-    ? possibleOriginals.file
-    : possibleOriginals.dir;
+  const originalLocation = stat.isFile ? possibleOriginals.file : possibleOriginals.dir;
 
-  if (await exists(possibleOriginals.file) || await exists(possibleOriginals.dir)) {
-    console.info(`"${originalLocation}" already exists. Keeping original and deleting "${location}"`);
-
-    await Deno.remove(location, { recursive: true })
-      .catch((e: unknown) => {
-        if (e instanceof Error && e.message.includes('another process')) throw new Error('You need to close the app before injecting');
-
-        throw e;
-      });
-  } else await Deno.rename(location, originalLocation)
+  if (await exists(possibleOriginals.file) || await exists(possibleOriginals.dir)) await Deno.remove(location, { recursive: true })
     .catch((e: unknown) => {
       if (e instanceof Error && e.message.includes('another process')) throw new Error('You need to close the app before injecting');
 
       throw e;
     });
+
+  else await Deno.rename(location, originalLocation)
+    .catch((e: unknown) => {
+      if (e instanceof Error && e.message.includes('another process')) throw new Error('You need to close the app before injecting');
+
+      throw e;
+    });
+}
+
+/**
+ * @param location The location to the original asar file/directory
+ * @param type The type of injection
+ */
+export async function injectLocal(location: string, type: 'packed' | 'unpacked' | 'symlink' = 'packed') {
+  if (!import.meta.dirname) throw new Error('This script must be run locally');
+
+  await setup(location);
 
   const popcornDist = type === 'packed'
     ? join(import.meta.dirname, '../../dist/electron.asar')
@@ -45,9 +49,7 @@ export async function injectLocal(location: string, type: 'packed' | 'unpacked' 
   if (!await exists(popcornDist)) {
     const { build } = await import('#build/index.mts');
 
-    await build('electron', {
-      dev: type === 'symlink',
-    });
+    await build('electron', { dev: type === 'symlink' });
   }
 
   if (type === 'symlink') {
@@ -60,7 +62,17 @@ export async function injectLocal(location: string, type: 'packed' | 'unpacked' 
   await copy(popcornDist, join(location, `../${basename(location).replace(/\.asar$/, '')}${type === 'packed' ? '.asar' : ''}`));
 }
 
+/**
+ * @param location The location to the original asar file/directory
+ * @param version The version of popcorn to inject
+ */
 export async function injectRemote(location: string, version: `v${string}`) {
-  const popcornAsarData = await fetch(`https://github.com/wathhr/popcorn/releases/${version}/download/electron.asar`).then(r => r.arrayBuffer());
+  await setup(location);
+
+  const popcornAsarData = await fetch(`https://github.com/wathhr/popcorn/releases/${version}/download/electron.asar`)
+    .then(r => r.arrayBuffer())
+    .catch((e) => {
+      throw new Error(`Something went wrong: ${e}`);
+    });
   await Deno.writeFile(join(location, `../${basename(location).replace('.asar', '')}.asar`), new Uint8Array(popcornAsarData));
 }
